+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Trust functionnality
+
+### Fixed
+- Handle UPnP port mapping removal exception
+- Don't print starting server when immediatly closing it
+
 ## [2.0.0] - 2020-05-01
 ### Added
 - Username functionnality
@@ -218,6 +226,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic server and client
 - Use regex to parse commands
 
+[Unreleased]: https://github.com/RomainL972/isn/compare/v2.0.0...HEAD
 [2.0.0]: https://github.com/RomainL972/isn/compare/v1.1.1...v2.0.0
 [1.1.1]: https://github.com/RomainL972/isn/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/RomainL972/isn/compare/v1.0.0...v1.1.0